Here's a build system. Creates a library suitable for static linking. It was really easy to make. Good for all POSIX systems (Linux, Mac. Tested on Linux.) but getting Windows to cooperate with the autotools is a hassle. Recommend setting up the MS toolchain for that.
TODO: Provide a sensible version number and Cliff's bug-report-address to AC_INIT.
TODO: Either use #defines in config.h, or fail out of configure if certain things aren't available.

It might be better for performance to compile & archive seperate objects, but things seem to be lightweight enough that it doesn't matter either way.